+++ example/src/app/(home)/components/index.tsx
@@ -88,7 +88,7 @@ const components: Component[] = [
   {
     title: 'Radio Group',
     icon: <ComponentIcon />,
-    path: 'radio',
+    path: 'radio-group',
   },
   {
     title: 'Scroll Shadow',
