+++ example/src/components/showcases/linear-task/dialogs/assignee.tsx
@@ -1,14 +1,7 @@
 import Feather from '@expo/vector-icons/Feather';
 import MaterialCommunityIcons from '@expo/vector-icons/MaterialCommunityIcons';
 import * as Haptics from 'expo-haptics';
-import {
-  Avatar,
-  Chip,
-  Dialog,
-  Radio,
-  RadioGroup,
-  useTheme,
-} from 'heroui-native';
+import { Avatar, Chip, Dialog, RadioGroup, useTheme } from 'heroui-native';
 import { useMemo, useState, type FC } from 'react';
 import { Platform, useWindowDimensions, View } from 'react-native';
 import { ScrollView } from 'react-native-gesture-handler';
@@ -155,7 +148,7 @@ export const Assignee: FC = () => {
                       className="self-stretch"
                       asChild
                     >
-                      <Radio
+                      <RadioGroup.Item
                         value={item.value}
                         onPress={() => {
                           if (Platform.OS === 'ios') {
@@ -165,25 +158,27 @@ export const Assignee: FC = () => {
                           }
                         }}
                       >
-                        <Radio.Content className="flex-row items-center gap-2">
+                        <RadioGroup.ItemContent className="flex-row items-center gap-2">
                           <View className="w-7 pl-0.5 justify-center">
                             <View className="scale-105">{item.indicator}</View>
                           </View>
-                          <Radio.Title>{item.label}</Radio.Title>
-                        </Radio.Content>
-                        <Radio.Indicator
+                          <RadioGroup.ItemTitle>
+                            {item.label}
+                          </RadioGroup.ItemTitle>
+                        </RadioGroup.ItemContent>
+                        <RadioGroup.ItemIndicator
                           colors={{
                             defaultBorder: 'transparent',
                             selectedBorder: 'transparent',
                           }}
                         >
-                          <Radio.IndicatorBackground
+                          <RadioGroup.ItemIndicatorBackground
                             colors={{
                               defaultBackground: 'transparent',
                               selectedBackground: 'transparent',
                             }}
                           />
-                          <Radio.IndicatorThumb>
+                          <RadioGroup.ItemIndicatorThumb>
                             {value === item.value && (
                               <Animated.View
                                 key={item.value}
@@ -196,9 +191,9 @@ export const Assignee: FC = () => {
                                 />
                               </Animated.View>
                             )}
-                          </Radio.IndicatorThumb>
-                        </Radio.Indicator>
-                      </Radio>
+                          </RadioGroup.ItemIndicatorThumb>
+                        </RadioGroup.ItemIndicator>
+                      </RadioGroup.Item>
                     </Dialog.Close>
                   ))}
                 </RadioGroup>
