+++ src/components/radio-group/index.ts
@@ -1,5 +1,18 @@
-export { default as RadioGroup, useRadioGroupContext } from './radio-group';
+export {
+  default as RadioGroup,
+  useRadioGroupContext,
+  useRadioGroupItemContext,
+} from './radio-group';
 export type {
   RadioGroupErrorMessageProps,
+  RadioGroupItemColor,
+  RadioGroupItemColors,
+  RadioGroupItemContentProps,
+  RadioGroupItemDescriptionProps,
+  RadioGroupItemIndicatorBackgroundProps,
+  RadioGroupItemIndicatorProps,
+  RadioGroupItemIndicatorThumbProps,
+  RadioGroupItemProps,
+  RadioGroupItemTitleProps,
   RadioGroupProps,
 } from './radio-group.types';
