--- src/components/radio/index.ts
@@ -1,11 +1,0 @@
-export { default as Radio, useRadioContext } from './radio';
-export type {
-  RadioColor,
-  RadioColors,
-  RadioDescriptionProps,
-  RadioIndicatorBackgroundProps,
-  RadioIndicatorProps,
-  RadioIndicatorThumbProps,
-  RadioProps,
-  RadioTitleProps,
-} from './radio.types';
